requirements.txt
numpy
pandas
transformers
torch
datasets
sentence-transformers
faiss-cpu
